pyproject.toml
[tool.poetry]
name = "crypto-sec-project"
version = "0.1.0"
description = ""
authors = ["Princekrampah <jsksprince@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
streamlit = "^1.35.0"
twilio = "^9.1.1"
python-dotenv = "^1.0.1"
python-decouple = "^3.8"
pycryptodome = "^3.20.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
